chore(standard-tests): hoist type utilities
